Reorganize uploader component divs
Previously, the file previews were inside the dropzone.  If drag and drop was not
enabled in the browser, would have had to recreate the preview div outside of dropzone.

Now, the dropzone is part of a div which encompases everything within the uploader div.

Also moved the browse button outside of the uploader component, so that it is visible
regardless of whether dropzone is enabled.

Before, files could not be dragged directly on top of the instruction text.
These changes also enable files to be dragged to any point within the dropzone.